refactor project level into crates
