+++ crates/game/Cargo.toml
@@ -1,0 +1,14 @@
+[package]
+name = "game"
+version = "0.1.0"
+edition = "2021"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+map_shared = {path = "../map_shared"}
+bevy = "0.10.1"
+serde = { version = "1.0", features = ["derive"] }
+serde_json = "1.0"
+bevy_egui = "0.20.2"
+egui_extras = "0.21.0"
