+++ crates/game/src/map/map_manager.rs
@@ -1,6 +1,9 @@
 use std::error::Error;
-use crate::map::json_types::MapData;
+use bevy::prelude::Resource;
 
+use map_shared::MapData;
+
+#[derive(Resource)]
 pub struct MapManager {
     pub tuto_map: MapData,
     pub level1_map: MapData,
@@ -10,13 +13,13 @@ pub struct MapManager {
 impl MapManager {
     pub fn new(tuto_map: MapData, level_map: MapData) -> Result<Self, Box<dyn Error>> {
         Ok(Self {
-            tuto_map: tuto_map,
+            tuto_map,
             level1_map: level_map,
             custom_map: None,
         })
     }
 
-    pub(crate) fn set_custom_map(&mut self, custom_map: super::json_types::MapData) {
+    pub fn set_custom_map(&mut self, custom_map: MapData) {
         self.custom_map = Some(custom_map)
     }
 }
