+++ crates/game/src/map/mod.rs
@@ -1,0 +1,2 @@
+pub mod map_manager;
+pub mod parser;
