+++ crates/game/src/plugins/menu/plugin.rs
