+++ crates/game/src/plugins/state/plugin.rs
