+++ crates/game/src/plugins/state/types.rs
