+++ crates/map_shared/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "the_veiled_path"
+name = "map_shared"
 version = "0.1.0"
 edition = "2021"
 
@@ -13,20 +13,9 @@ lto = "thin"
 codegen-units = 1
 strip = true
 
-[workspace]
-members = [
-    "crates/*",
-]
-
-
 [dependencies]
-game = {path = "crates/game"}
 bevy = "0.10.1"
 serde = { version = "1.0", features = ["derive"] }
 serde_json = "1.0"
 bevy_egui = "0.20.2"
 egui_extras = "0.21.0"
-
-[dev-dependencies]
-ldtk_rust = "0.6"
-map_shared = {path = "crates/map_shared"}
