--- src/map/mod.rs
@@ -1,3 +1,0 @@
-pub(crate) mod json_types;
-pub(crate) mod map_manager;
-pub(crate) mod parser;
